# docs/jobhub/03-openapi.yaml
openapi: 3.0.3
info:
  title: JobHub V0.1 API
  version: 0.1.0
  description: |
    本地单用户模式 API。状态转换以 `02-state-machines.md` 为准；未经用户确认的候选要求不得作为确定性结论。
    所有 POST、PUT、DELETE 及状态转换接口均应支持 Idempotency-Key；带 version 的资源更新应提交 If-Match-Version。
servers:
  - url: http://localhost:8080/api
    description: 本地开发服务
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: 查询岗位列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: decisionStatus
          in: query
          schema: { $ref: '#/components/schemas/JobDecisionStatus' }
        - name: jobStatus
          in: query
          schema: { $ref: '#/components/schemas/JobStatus' }
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: 岗位分页列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageJob' }
    post:
      tags: [Jobs]
      summary: 创建岗位并保存 JD 原文
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreateRequest' }
      responses:
        '201':
          description: 创建成功
          content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
  /jobs/{jobId}:
    parameters: [ { $ref: '#/components/parameters/JobId' } ]
    get:
      tags: [Jobs]
      summary: 获取岗位详情
      responses:
        '200': { description: 岗位详情, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Jobs]
      summary: 更新岗位基础信息或投递决定
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobUpdateRequest' }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/BusinessRuleError' }
  /jobs/{jobId}/archive:
    parameters: [ { $ref: '#/components/parameters/JobId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
    post:
      tags: [Jobs]
      summary: 归档岗位
      responses:
        '200': { description: 已归档, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '422': { $ref: '#/components/responses/BusinessRuleError' }
  /jobs/{jobId}/restore:
    parameters: [ { $ref: '#/components/parameters/JobId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
    post:
      tags: [Jobs]
      summary: 恢复已归档岗位
      responses:
        '200': { description: 已恢复, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
  /jobs/{jobId}/requirements/extract:
    parameters: [ { $ref: '#/components/parameters/JobId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
    post:
      tags: [Jobs]
      summary: 使用基础规则提取候选岗位要求
      description: P0 同步规则提取；不使用 AI，结果均为 PENDING，且不会自动覆盖已确认要求。
      responses:
        '200':
          description: 候选要求
          content: { application/json: { schema: { $ref: '#/components/schemas/RequirementExtractionResult' } } }
        '422': { $ref: '#/components/responses/BusinessRuleError' }
  /jobs/{jobId}/requirements:
    parameters: [ { $ref: '#/components/parameters/JobId' } ]
    get:
      tags: [Jobs]
      summary: 查询岗位要求及确认状态
      responses:
        '200':
          description: 岗位要求列表
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/JobRequirement' } } } }
  /job-requirements/{requirementId}:
    parameters: [ { $ref: '#/components/parameters/RequirementId' } ]
    put:
      tags: [Jobs]
      summary: 确认、编辑、忽略或恢复单项岗位要求
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequirementUpdateRequest' }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/JobRequirement' } } } }
        '409': { $ref: '#/components/responses/VersionConflict' }
...
  /jobs/{jobId}/gap-list:
    parameters: [ { $ref: '#/components/parameters/JobId' } ]
    get:
      tags: [Jobs]
      summary: 获取可解释差距清单
      responses:
        '200':
          description: 差距项；只基于 CONFIRMED 要求
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/GapItem' } } } }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema: { type: integer, minimum: 0 }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    JobId: { name: jobId, in: path, required: true, schema: { type: string, format: uuid } }
    RequirementId: { name: requirementId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    NotFound:
      description: 资源不存在或不可见
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError:
      description: 请求字段校验失败
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    VersionConflict:
      description: 版本冲突
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    IdempotencyConflict:
      description: 幂等键与既有请求不一致
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BusinessRuleError:
      description: 业务规则不满足
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string, example: ILLEGAL_STATE_TRANSITION }
        message: { type: string, example: 当前状态不允许转换为目标状态 }
        traceId: { type: string }
        fieldErrors: { type: array, items: { $ref: '#/components/schemas/FieldError' } }
        currentState: { type: string }
        targetState: { type: string }
        reason: { type: string }
    FieldError:
      type: object
      required: [field, message]
      properties: { field: { type: string }, message: { type: string } }
...
    JobStatus: { type: string, enum: [ACTIVE, ARCHIVED] }
    JobDecisionStatus: { type: string, nullable: true, enum: [TO_APPLY, APPLY, DEFER, IGNORE] }
    RequirementConfirmationStatus: { type: string, enum: [PENDING, CONFIRMED, IGNORED] }
    RequirementType: { type: string, enum: [MUST, BONUS, RESPONSIBILITY, EXPERIENCE, DOMAIN, TO_CONFIRM] }
    GapStatus: { type: string, enum: [SATISFIED_WITH_EVIDENCE, SELF_REPORTED_NO_EVIDENCE, NOT_MET, INSUFFICIENT_INFO, PENDING_CONFIRMATION] }
...
    JobCreateRequest:
      type: object
      required: [companyName, title, jdRawText]
      properties:
        companyName: { type: string, maxLength: 100 }
        title: { type: string, maxLength: 150 }
        jdRawText: { type: string, minLength: 20, maxLength: 50000 }
        source: { type: string, maxLength: 100 }
        sourceUrl: { type: string, format: uri, nullable: true }
        location: { type: string, maxLength: 100 }
        salaryRange: { type: string, maxLength: 100 }
        notes: { type: string, maxLength: 5000 }
    JobUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/JobCreateRequest'
        - type: object
          properties:
            decisionStatus: { $ref: '#/components/schemas/JobDecisionStatus' }
            decisionReason: { type: string, maxLength: 1000, nullable: true }
    Job:
      type: object
      required: [id, companyName, title, jdRawText, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        companyName: { type: string }
        title: { type: string }
        jdRawText: { type: string }
        source: { type: string, nullable: true }
        sourceUrl: { type: string, format: uri, nullable: true }
        location: { type: string, nullable: true }
        salaryRange: { type: string, nullable: true }
        decisionStatus: { $ref: '#/components/schemas/JobDecisionStatus' }
        decisionReason: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/JobStatus' }
        notes: { type: string, nullable: true }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobRequirement:
      type: object
      required: [id, jobId, rawText, type, confirmationStatus, version]
      properties:
        id: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        rawText: { type: string }
        normalizedName: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/RequirementType' }
        proficiencyText: { type: string, nullable: true }
        confirmationStatus: { $ref: '#/components/schemas/RequirementConfirmationStatus' }
        source: { type: string, enum: [RULE, USER, AI], default: RULE }
        version: { type: integer }
    RequirementUpdateRequest:
      type: object
      required: [confirmationStatus]
      properties:
        rawText: { type: string, maxLength: 2000 }
        normalizedName: { type: string, maxLength: 200 }
        type: { $ref: '#/components/schemas/RequirementType' }
        proficiencyText: { type: string, maxLength: 500 }
        confirmationStatus: { $ref: '#/components/schemas/RequirementConfirmationStatus' }
        reason: { type: string, maxLength: 1000, description: 人工修正匹配状态时的原因；与 manualMatchStatus 配合使用。 }
        manualMatchStatus:
          $ref: '#/components/schemas/GapStatus'
          description: 可选。用户提供后，对应要求的 requirement_match 将被人工修正为此状态；reason 字段作为修正原因保留。原计算依据通过 evidence_snapshot_json 保留（AT-04）。
...
    RequirementExtractionResult:
      type: object
      required: [jobId, extractedAt, candidates]
      properties:
        jobId: { type: string, format: uuid }
        extractedAt: { type: string, format: date-time }
        candidates: { type: array, items: { $ref: '#/components/schemas/JobRequirement' } }
    GapItem:
      type: object
      required: [requirement, status, evidence]
      properties:
        requirement: { $ref: '#/components/schemas/JobRequirement' }
        status: { $ref: '#/components/schemas/GapStatus' }
        evidence:
          type: array
          description: 用于本项结论的用户事实；可为空。
          items: { $ref: '#/components/schemas/EvidenceReference' }
        manualOverrideReason: { type: string, nullable: true }
...
    EvidenceReference:
      type: object
      required: [id, type, title]
      properties: { id: { type: string, format: uuid }, type: { type: string }, title: { type: string }, urlOrPath: { type: string, nullable: true } }
...
    PageJob:
      type: object
      required: [items, page, pageSize, total]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Job' } }, page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }
...